Rename Input to LimitInput
