+++ frontend/src/metabase/query_builder/components/notebook/steps/LimitStep.jsx
@@ -3,9 +3,8 @@ import React from "react";
 
 import { t } from "ttag";
 
-import Input from "metabase/components/Input";
-
 import { NotebookCell } from "../NotebookCell";
+import LimitInput from "metabase/query_builder/components/LimitInput";
 
 export default function LimitStep({
   color,
@@ -16,7 +15,7 @@ export default function LimitStep({
 }) {
   return (
     <NotebookCell color={color}>
-      <Input
+      <LimitInput
         small
         className="mb1"
         type="number"
